added login, logout, authentication

+++ blog/views.py
@@ -5,11 +5,11 @@ from database import session
 from models import Post
 
 import mistune
-from flask import request, redirect, url_for
+from flask import request, redirect, url_for, g
 
 # imports for login and creating users
 from flask import flash
-from flask.ext.login import login_user
+from flask.ext.login import login_user, logout_user
 from werkzeug.security import check_password_hash
 from models import User
 
@@ -47,23 +47,20 @@ def posts(page=1, paginate_by=10):
     )
 
 
-@app.route("/post/add", methods=["GET"])
+@app.route("/post/add", methods=["GET", "POST"])
 @login_required
-def add_post_get():
-    return render_template("add_post.html")
-
-
-@app.route("/post/add", methods=["POST"])
-@login_required
-def add_post_post():
-    post = Post(
+def add_post():
+    if request.method == "POST":
+        post = Post(
         title=request.form["title"],
         content=mistune.markdown(request.form["content"]),
         author=current_user
-    )
-    session.add(post)
-    session.commit()
-    return redirect(url_for("posts"))
+        )
+        session.add(post)
+        session.commit()
+        return redirect(url_for("posts"))
+    
+    return render_template("add_post.html")
 
 
 @app.route("/post/<int:id>")
@@ -72,33 +69,33 @@ def view_post(id):
     return render_template("view_post.html", post=post)
 
 
-@app.route("/post/<int:id>/edit", methods=["GET"])
+@app.route("/post/<int:id>/edit", methods=["GET", "POST"])
 @login_required
-def edit_post_get(id):
+def edit_post(id):
+    if request.method == "POST":
+        post = session.query(Post).get(id)
+        post.title = request.form["title"]
+        post.content = mistune.markdown(request.form["content"])
+        session.add(post)
+        session.commit()
+        return redirect(url_for("posts"))
+
     post = session.query(Post).get(id)
     return render_template("edit_post.html", post=post)
 
 
-@app.route("/post/<int:id>/edit", methods=["POST"])
-@login_required
-def edit_post_post(id):
-    post = session.query(Post).get(id)
-    post.title = request.form["title"]
-    post.content = mistune.markdown(request.form["content"])
-    session.add(post)
-    session.commit()
-    return redirect(url_for("posts"))
-
-
 @app.route("/post/<int:id>/delete", methods=["GET", "POST"])
 @login_required
-def delete_post_get(id):
+def delete_post(id):
     if request.method == "POST":
         post = session.query(Post).get(id)
-        session.delete(post)
-        session.commit()
-        flash("Message deleted!", "success")
+        if post.author_id == g.user:
+            session.delete(post)
+            session.commit()
+            flash("Message deleted!", "success")
+        flash("Sorry, you cannot delete that", "danger")
         return redirect(url_for("posts"))
+        
 
     post = session.query(Post).get(id)
     return render_template("delete_post.html", post=post)
@@ -120,3 +117,14 @@ def login_post():
 
     login_user(user)
     return redirect(request.args.get('next') or url_for("posts"))
+
+
+@app.route("/logout")
+def logout():
+    logout_user()
+    flash("You have been logged out", "danger")
+    return redirect(url_for('posts'))
+
+@app.before_request
+def before_request():
+    g.user = current_user
